requirements.txt
alembic==1.4.3
astroid==2.4.2
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dill==0.3.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
klepto==0.2.0
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pox==0.2.9
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
pykakasi==2.0.6
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.2
traitlets==5.0.5
tweepy==3.10.0
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
